macd_vol_signal_bot/requirements.txt
ccxt>=4.5.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
pyyaml>=6.0.0
openai>=1.0.0
